refactor: create pointercrate-user config.rs
Move the `secret()` function from pointercrate-core to
pointercrate-user, since the secret key is only used in
pointercrate-user. Also opens us up for having more pointercrate-user
specific configuration options (such as oauth client id).
